Updated the StoreProduct, DBStoreProduct, and DBStoreProductQuery to use Timestamp instead of Date.
Implemented searchTitle in DBItemQuery. This method will search the titles
given a particular title. It will return all of the titles that are
similar to the given String from most likely to least.

Added a getRecentDate to PriceHistoryQuery and DBPriceHistoryQuery. This
method will get the most recent date for the given store and item.